# Refactor the spacepoint formation