--- a/core/include/traccc/seeding/detail/spacepoint_formation.hpp
+++ b/core/include/traccc/seeding/detail/spacepoint_formation.hpp
@@ -11,21 +11,19 @@
 #include "traccc/definitions/qualifiers.hpp"
 #include "traccc/edm/measurement.hpp"
 #include "traccc/edm/spacepoint.hpp"
-#include "traccc/geometry/silicon_detector_description.hpp"
 
 namespace traccc::details {
 
 /// Function helping with filling/setting up a spacepoint object
 ///
-/// @param[out] sp          The spacepoint to fill / set up
+/// @param[in]  det         The tracking geometry
 /// @param[in]  measurement The measurement to create the spacepoint out of
-/// @param[in]  dd          The detector description
 ///
-TRACCC_HOST_DEVICE inline void fill_spacepoint(
-    spacepoint& sp, const measurement& meas,
-    const silicon_detector_description::const_device& dd);
+template <typename detector_t>
+TRACCC_HOST_DEVICE inline spacepoint create_spacepoint(const detector_t& det,
+                                                       const measurement& meas);
 
 }  // namespace traccc::details
 
 // Include the implementation.
-#include "traccc/clusterization/impl/spacepoint_formation.ipp"
+#include "traccc/seeding/impl/spacepoint_formation.ipp"