--- a/device/common/include/traccc/seeding/device/impl/form_spacepoints.ipp
+++ b/device/common/include/traccc/seeding/device/impl/form_spacepoints.ipp
@@ -8,18 +8,17 @@
 #pragma once
 
 // Project include(s).
-#include "traccc/clusterization/details/spacepoint_formation.hpp"
+#include "traccc/seeding/detail/spacepoint_formation.hpp"
 
 // System include(s).
 #include <cassert>
 
 namespace traccc::device {
 
-TRACCC_HOST_DEVICE
-inline void form_spacepoints(
-    std::size_t globalIndex,
+template <typename detector_t>
+TRACCC_HOST_DEVICE inline void form_spacepoints(
+    std::size_t globalIndex, typename detector_t::view_type det_view,
     const measurement_collection_types::const_view& measurements_view,
-    const silicon_detector_description::const_view& det_descr_view,
     unsigned int measurement_count,
     spacepoint_collection_types::view spacepoints_view) {
 
@@ -28,17 +27,19 @@ inline void form_spacepoints(
         return;
     }
 
+    // Create the tracking geometry
+    detector_t det(det_view);
+
     // Get device copy of input/output parameters
     const measurement_collection_types::const_device measurements(
         measurements_view);
     assert(measurements.size() == measurement_count);
-    const silicon_detector_description::const_device det_descr(det_descr_view);
     spacepoint_collection_types::device spacepoints(spacepoints_view);
-    assert(spacepoints.size() == measurement_count);
+
+    const auto& meas = measurements.at(globalIndex);
 
     // Fill the spacepoint using the common function.
-    details::fill_spacepoint(spacepoints.at(globalIndex),
-                             measurements.at(globalIndex), det_descr);
+    spacepoints.push_back(details::create_spacepoint(det, meas));
 }
 
 }  // namespace traccc::device