--- a/device/cuda/include/traccc/cuda/seeding/spacepoint_formation_algorithm.hpp
+++ b/device/cuda/include/traccc/cuda/seeding/spacepoint_formation_algorithm.hpp
@@ -7,11 +7,12 @@
 
 #pragma once
 
+// Local include(s).
+#include "traccc/cuda/utils/stream.hpp"
+
 // Project include(s).
-#include "traccc/edm/cell.hpp"
 #include "traccc/edm/measurement.hpp"
 #include "traccc/edm/spacepoint.hpp"
-#include "traccc/geometry/silicon_detector_description.hpp"
 #include "traccc/utils/algorithm.hpp"
 #include "traccc/utils/memory_resource.hpp"
 
@@ -21,17 +22,18 @@
 // System include(s).
 #include <functional>
 
-namespace traccc::alpaka {
+namespace traccc::cuda {
 
 /// Algorithm forming space points out of measurements
 ///
 /// This algorithm performs the local-to-global transformation of the 2D
 /// measurements made on every detector module, into 3D spacepoint coordinates.
 ///
+template <typename detector_t>
 class spacepoint_formation_algorithm
     : public algorithm<spacepoint_collection_types::buffer(
-          const measurement_collection_types::const_view&,
-          const silicon_detector_description::const_view&)> {
+          const typename detector_t::view_type&,
+          const measurement_collection_types::const_view&)> {
 
     public:
     /// Constructor for spacepoint_formation
@@ -39,19 +41,19 @@ class spacepoint_formation_algorithm
     /// @param mr is the memory resource
     ///
     spacepoint_formation_algorithm(const traccc::memory_resource& mr,
-                                   vecmem::copy& copy);
+                                   vecmem::copy& copy, stream& str);
 
     /// Callable operator for the space point formation, based on one single
     /// module
     ///
-    /// @param measurements_view A collection of measurements
-    /// @param det_descr         The detector description
+    /// @param det_view     A view type object of tracking geometry
+    /// @param measurements All reconstructed measurements in an event
     /// @return A spacepoint container, with one spacepoint for every
     ///         measurement
     ///
-    output_type operator()(
-        const measurement_collection_types::const_view& measurements_view,
-        const silicon_detector_description::const_view& det_descr)
+    spacepoint_collection_types::buffer operator()(
+        const typename detector_t::view_type& det_view,
+        const measurement_collection_types::const_view& measurements_view)
         const override;
 
     private:
@@ -59,6 +61,9 @@ class spacepoint_formation_algorithm
     traccc::memory_resource m_mr;
     /// The copy object to use
     std::reference_wrapper<vecmem::copy> m_copy;
+    /// The CUDA stream to use
+    std::reference_wrapper<stream> m_stream;
+
 };  // class spacepoint_formation_algorithm
 
-}  // namespace traccc::alpaka
+}  // namespace traccc::cuda