… core/CMakeLists.txt
… core/include/traccc/clusterization/impl/spacepoint_formation.ipp
… core/include/traccc/seeding/experimental/spacepoint_formation.hpp
… core/include/traccc/seeding/experimental/spacepoint_formation.ipp
… device/common/CMakeLists.txt
… device/common/include/traccc/seeding/device/experimental/form_spacepoints.hpp
… device/common/include/traccc/seeding/device/experimental/impl/form_spacepoints.ipp
… device/cuda/CMakeLists.txt
… device/cuda/include/traccc/cuda/clusterization/spacepoint_formation_algorithm.hpp
… device/cuda/src/seeding/experimental/spacepoint_formation.cu
… device/sycl/CMakeLists.txt
… device/sycl/include/traccc/sycl/clusterization/spacepoint_formation_algorithm.hpp
… device/sycl/src/clusterization/spacepoint_formation_algorithm.sycl
… device/sycl/src/seeding/experimental/spacepoint_formation.sycl
… examples/run/alpaka/seq_example_alpaka.cpp
… examples/run/cpu/full_chain_algorithm.cpp
… examples/run/cpu/full_chain_algorithm.hpp
… examples/run/cpu/seq_example.cpp
… examples/run/cuda/full_chain_algorithm.cpp
… examples/run/cuda/full_chain_algorithm.hpp
… examples/run/cuda/seq_example_cuda.cpp
… examples/run/openmp/par_example.cpp
… examples/run/sycl/full_chain_algorithm.hpp
… examples/run/sycl/full_chain_algorithm.sycl
… examples/run/sycl/seq_example_sycl.sycl
… tests/cpu/test_clusterization_resolution.cpp
… tests/cpu/test_spacepoint_formation.cpp
… tests/cuda/test_spacepoint_formation.cpp
… tests/sycl/test_spacepoint_formation.sycl