This are the changes that I have been suggesting for long time.

1. `spacepoint_formation` takes detray detector instead of detector description. Direct use of `point_to_global` with `transform_matrix` should be avoided. Most importantly, the local-to-global transformation of annulus shape (Polar coordinate) does not work with `point_to_global` which only works for Cartesian coordinate.

2. As detector description is not required (and should not be used), I moved its directory to `seeding` where the `experimental::speacepoing_formation` is located. This makes sense because spacepoints are technically only used for seeding. 

~~+ Spacepoint formation is only applied to the 2D measurements as we do not support 1D measurements sp formation. Supporting 1D measurement will need to be followed up by further studies~~

Let me have 2D condition in the next PR (to test the performance without changing the actual results...)