@beomki-yeo 2024-09-30
The performance remains the same

Main

```
Reconstructed track parameters: 3299715
Time totals:
                  File reading  617 ms
            Warm-up processing  147130 ms
              Event processing  4809 ms
Throughput:
            Warm-up processing  14713 ms/event, 0.067967 events/s
              Event processing  48.0923 ms/event, 20.7934 events/s
```

PR

```
Reconstructed track parameters: 3299774
Time totals:
                  File reading  618 ms
            Warm-up processing  938 ms
              Event processing  4811 ms
Throughput:
            Warm-up processing  93.8457 ms/event, 10.6558 events/s
              Event processing  48.1174 ms/event, 20.7825 events/s
```